+++ src/modules/caen/CaenBase.h
@@ -8,7 +8,7 @@
 //===----------------------------------------------------------------------===//
 #pragma once
 
-#include <caen/Counters.h>
+#include <caen/CaenCounters.h>
 #include <common/detector/Detector.h>
 #include <common/kafka/EV44Serializer.h>
