+++ src/modules/caen/CaenInstrument.h
@@ -13,7 +13,7 @@
 
 #include <bifrost/geometry/BifrostGeometry.h>
 #include <caen/CaenBase.h> // to get CaenSettings
-#include <caen/Counters.h>
+#include <caen/CaenCounters.h>
 #include <caen/geometry/Config.h>
 #include <caen/readout/Readout.h>
 #include <common/readout/ess/ESSTime.h>
